Completed Sound System
implimented the sound system using the FMOD library, allowing us to be
able to read in sound files of mutiple types, apply them to SoundSystem
objects and excecute playback, as well as a few other helpful functions.
Edited the SceneObjects to hold a SoundSystem on them, or nullptr if not
wanted. Also implimented the Audiomanager and connected that to the
resource manager so audio files will be stored properly.